+++ NEWS.md
@@ -3,6 +3,11 @@
 -   Fixed a typo in the lookup path that meant the package wouldn't work on Windows (RStudio Desktop).
 -   Make the minimum required R version 4.1
 -   Overhaul the README; fixed out of date information, and added more detail and examples.
+-   Metadata is now available! Currently, this is only for the Scottish Postcode Directory (`get_spd()`), but we will bring it to the other lookups soon, too. You can see the metadata by using `metadata()` on the lookup object, for example:
+```r
+spd <- get_spd()
+metadata(spd)
+```
 
 # phslookups 0.1.0